+++ edgy/contrib/contenttypes/models.py
@@ -24,7 +24,11 @@ class ContentType(edgy.Model, metaclass=ContentTypeMeta):
 
     async def get_instance(self) -> edgy.Model:
         reverse_name = f"reverse_{self.name.lower()}"
-        return await cast("QuerySet", getattr(self, reverse_name)).using(self.schema_name).get()
+        return (
+            await cast("QuerySet", getattr(self, reverse_name))
+            .using(schema=self.schema_name)
+            .get()
+        )
 
     async def delete(
         self, skip_post_delete_hooks: bool = False, remove_referenced_call: bool = False
@@ -33,4 +37,4 @@ class ContentType(edgy.Model, metaclass=ContentTypeMeta):
         query = cast("QuerySet", getattr(self, reverse_name))
         await super().delete(skip_post_delete_hooks=skip_post_delete_hooks)
         if not remove_referenced_call and self.no_constraint:
-            await query.using(self.schema_name).delete()
+            await query.using(schema=self.schema_name).delete()