+++ tests/contrib/multi_tenancy/test_tenant_models_working.py
@@ -62,9 +62,9 @@ async def test_schema():
     )
 
     for i in range(5):
-        await Product.query.using(tenant.schema_name).create(name=f"product-{i}")
+        await Product.query.using(schema=tenant.schema_name).create(name=f"product-{i}")
 
-    total = await Product.query.using(tenant.schema_name).all()
+    total = await Product.query.using(schema=tenant.schema_name).all()
 
     assert len(total) == 5
 
@@ -79,7 +79,7 @@ async def test_schema():
 
     assert len(total) == 15
 
-    total = await Product.query.using(tenant.schema_name).all()
+    total = await Product.query.using(schema=tenant.schema_name).all()
 
     assert len(total) == 5
 
@@ -93,35 +93,37 @@ async def test_can_have_multiple_tenants_with_different_records():
     )
 
     # Create a user for edgy
-    user_edgy = await User.query.using(edgy.schema_name).create(name="Edgy")
+    user_edgy = await User.query.using(schema=edgy.schema_name).create(name="Edgy")
 
     # Create products for user_edgy
     for i in range(5):
-        await Product.query.using(edgy.schema_name).create(name=f"product-{i}", user=user_edgy)
+        await Product.query.using(schema=edgy.schema_name).create(
+            name=f"product-{i}", user=user_edgy
+        )
 
     # Create a user for saffier
-    user_saffier = await User.query.using(saffier.schema_name).create(name="Saffier")
+    user_saffier = await User.query.using(schema=saffier.schema_name).create(name="Saffier")
 
     # Create products for user_saffier
     for i in range(25):
-        await Product.query.using(saffier.schema_name).create(
+        await Product.query.using(schema=saffier.schema_name).create(
             name=f"product-{i}", user=user_saffier
         )
 
     # Create top level users
     for name in range(10):
-        await User.query.using(saffier.schema_name).create(name=f"user-{name}")
-        await User.query.using(edgy.schema_name).create(name=f"user-{name}")
+        await User.query.using(schema=saffier.schema_name).create(name=f"user-{name}")
+        await User.query.using(schema=edgy.schema_name).create(name=f"user-{name}")
         await User.query.create(name=f"user-{name}")
 
     # Check the totals
     top_level_users = await User.query.all()
     assert len(top_level_users) == 10
 
-    users_edgy = await User.query.using(edgy.schema_name).all()
+    users_edgy = await User.query.using(schema=edgy.schema_name).all()
     assert len(users_edgy) == 11
 
-    users_saffier = await User.query.using(saffier.schema_name).all()
+    users_saffier = await User.query.using(schema=saffier.schema_name).all()
     assert len(users_saffier) == 11
 
 
@@ -130,20 +132,20 @@ async def test_model_crud():
         schema_name="edgy", domain_url="https://edgy.tarsild.io", tenant_name="edgy"
     )
 
-    users = await User.query.using(edgy.schema_name).all()
+    users = await User.query.using(schema=edgy.schema_name).all()
     assert users == []
 
-    user = await User.query.using(edgy.schema_name).create(name="Test")
-    users = await User.query.using(edgy.schema_name).all()
+    user = await User.query.using(schema=edgy.schema_name).create(name="Test")
+    users = await User.query.using(schema=edgy.schema_name).all()
     assert user.name == "Test"
     assert user.pk is not None
     assert users == [user]
 
-    lookup = await User.query.using(edgy.schema_name).get()
+    lookup = await User.query.using(schema=edgy.schema_name).get()
     assert lookup == user
 
     await user.update(name="Jane")
-    users = await User.query.using(edgy.schema_name).all()
+    users = await User.query.using(schema=edgy.schema_name).all()
     assert user.name == "Jane"
     assert user.pk is not None
     assert users == [user]
@@ -153,5 +155,5 @@ async def test_model_crud():
     assert users == []
 
     await user.delete()
-    users = await User.query.using(edgy.schema_name).all()
+    users = await User.query.using(schema=edgy.schema_name).all()
     assert users == []